debug

+++ turn.m
@@ -1,6 +1,6 @@
 function turn(angle)
     TurningSpeed    = 50;
-    turnTicks       = uint16((300/45)*(angle/2));      % assuming 45dgs turn is 219 ticks
+    turnTicks       = int16((300/45)*(angle/2));      % assuming 45dgs turn is 219 ticks
     Ports           = [MOTOR_B; MOTOR_C];  % motorports for left and right wheel
     
     mTurn1                      = NXTMotor(Ports(2)); % is it needed to swap ports?
@@ -10,9 +10,9 @@ function turn(angle)
     
     % where are we?
     mTurn1.ResetPosition();
-    data    = mTurn1.ReadFromNXT();
-    pos     = data.Position;
-    mTurn1.TachoLimit           = turnTicks + pos;
+    data                        = mTurn1.ReadFromNXT();
+    pos                         = data.Position;
+    mTurn1.TachoLimit           =  int16(turnTicks + pos);
     
     
     mTurn2          = mTurn1;
@@ -26,7 +26,7 @@ function turn(angle)
     mTurn2.ResetPosition();
     data    = mTurn2.ReadFromNXT();
     pos     = data.Position;
-    mTurn2.TachoLimit           = turnTicks + pos;
+    mTurn2.TachoLimit           =  int16(turnTicks + pos);
     
     mTurn2.SendToNXT(); % then turn angle/2 by driving the othew wheel backward
     mTurn2.WaitFor();
